+++ apps/backoffice-v2/src/pages/Entity/components/Case/Case.Documents.tsx
@@ -84,7 +84,7 @@ export const Documents: FunctionComponent<IDocumentsProps> = ({ documents, isLoa
             </TransformComponent>
           </TransformWrapper>
           <div className={`absolute z-50 flex space-x-2 bottom-right-6`}>
-            {!isPdf(selectedImage) && (
+            {!isPdf(selectedImage) && !isLoading && (
               <>
                 <button
                   type={`button`}
@@ -95,40 +95,40 @@ export const Documents: FunctionComponent<IDocumentsProps> = ({ documents, isLoa
                 >
                   <FileText className={`rotate-90 p-0.5`} />
                 </button>
-                <button
-                  className={ctw(
-                    'btn btn-circle btn-ghost btn-sm bg-base-300/70 focus:outline-primary',
-                    {
-                      hidden: !isCropping,
-                    },
-                  )}
-                  onClick={onCancelCrop}
-                >
-                  <XMarkSvg className={`p-0.5`} />
-                </button>
-                <div
-                  title={
-                    isRotatedOrTransformed
-                      ? `Cannot OCR rotated, zoomed, panned, or pinched documents`
-                      : undefined
-                  }
-                >
-                  <button
-                    type={`button`}
-                    className={ctw(
-                      `btn btn-circle btn-ghost btn-sm bg-base-300/70 text-[0.688rem] focus:outline-primary`,
-                      { loading: isLoadingOCR },
-                    )}
-                    onClick={onOcr}
-                    disabled={isLoading || isRotatedOrTransformed}
-                  >
-                    {isCropping && !isLoadingOCR && <CheckSvg className={`p-0.5`} />}
-                    {!isCropping && !isLoadingOCR && <span className={`p-0.5`}>OCR</span>}
-                  </button>
-                </div>
+                {/*<button*/}
+                {/*  className={ctw(*/}
+                {/*    'btn btn-circle btn-ghost btn-sm bg-base-300/70 focus:outline-primary',*/}
+                {/*    {*/}
+                {/*      hidden: !isCropping,*/}
+                {/*    },*/}
+                {/*  )}*/}
+                {/*  onClick={onCancelCrop}*/}
+                {/*>*/}
+                {/*  <XMarkSvg className={`p-0.5`} />*/}
+                {/*</button>*/}
+                {/*<div*/}
+                {/*  title={*/}
+                {/*    isRotatedOrTransformed*/}
+                {/*      ? `Cannot OCR rotated, zoomed, panned, or pinched documents`*/}
+                {/*      : undefined*/}
+                {/*  }*/}
+                {/*>*/}
+                {/*  <button*/}
+                {/*    type={`button`}*/}
+                {/*    className={ctw(*/}
+                {/*      `btn btn-circle btn-ghost btn-sm bg-base-300/70 text-[0.688rem] focus:outline-primary`,*/}
+                {/*      { loading: isLoadingOCR },*/}
+                {/*    )}*/}
+                {/*    onClick={onOcr}*/}
+                {/*    disabled={isRotatedOrTransformed}*/}
+                {/*  >*/}
+                {/*    {isCropping && !isLoadingOCR && <CheckSvg className={`p-0.5`} />}*/}
+                {/*    {!isCropping && !isLoadingOCR && <span className={`p-0.5`}>OCR</span>}*/}
+                {/*  </button>*/}
+                {/*</div>*/}
               </>
             )}
-            <ImageViewer.ZoomButton disabled={isLoading} />
+            {!isLoading && <ImageViewer.ZoomButton />}
           </div>
         </div>
       </div>