+++ examples/kyb-app/src/components/providers/CustomerProvider/types.ts
@@ -1,4 +1,4 @@
-import { TCustomer } from '@app/domains/customer';
+import { TCustomer } from '@app/domains/collection-flow';
 
 export interface CustomerContext {
   customer: TCustomer;