+++ services/workflows-service/src/collection-flow/dto/update-flow-input.dto.ts
@@ -113,9 +113,6 @@ export class UpdateFlowPayload {
 }
 
 export class UpdateFlowDto {
-  @IsString()
-  flowType!: string;
-
   @ValidateNested()
   @IsNotEmpty()
   @Type(() => UpdateFlowPayload)